% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-panel.R
\name{check_panel_cols}
\alias{check_panel_cols}
\title{Validate required columns in panel data}
\usage{
check_panel_cols(dta, id_col, time_col, tr_col)
}
\arguments{
\item{dta}{data.table}

\item{id_col}{unit id column name}

\item{time_col}{time column name}

\item{tr_col}{treatment column name}
}
\description{
Validate required columns in panel data
}
