% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampling.R
\name{.compute_wss}
\alias{.compute_wss}
\title{Weighted sum of squared residuals (for sigma2 update)}
\usage{
.compute_wss(X_list, w_list, y_list, beta_bd, K, J0, nu_list = NULL, rho = 0)
}
\description{
Weighted sum of squared residuals (for sigma2 update)
}
\keyword{internal}
