% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampling.R
\name{.compute_XtWy_units}
\alias{.compute_XtWy_units}
\title{Compute XtWy per unit for current y_tilde}
\usage{
.compute_XtWy_units(X_list, w_list, y_tilde_list, K, J0)
}
\description{
Compute XtWy per unit for current y_tilde
}
\keyword{internal}
