% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampling.R
\name{.sample_z_star}
\alias{.sample_z_star}
\title{Fast truncated normal sampler (inverse-CDF, no ifelse dispatch)}
\usage{
.sample_z_star(eta, d_vec, treat_idx, ctrl_idx)
}
\description{
Fast truncated normal sampler (inverse-CDF, no ifelse dispatch)
}
\keyword{internal}
