% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampling.R
\name{gibbs_postscm}
\alias{gibbs_postscm}
\title{Gibbs sampler dispatcher}
\usage{
gibbs_postscm(gdata, n_iter = 1000, burn_in = 500, control = NULL)
}
\arguments{
\item{gdata}{Output from prepare_data_general().}

\item{n_iter}{Integer. Total Gibbs iterations.}

\item{burn_in}{Integer. Burn-in iterations discarded.}

\item{control}{Optional list of prior/control parameters.}
}
\value{
List of posterior draw matrices.
}
\description{
Gibbs sampler dispatcher
}
