% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-math.R
\name{initialize_dense_covariance}
\alias{initialize_dense_covariance}
\title{Initialize a dense covariance matrix for LZ endogenous variables}
\usage{
initialize_dense_covariance(LZ, diag_value = 1)
}
\arguments{
\item{LZ}{Integer number of endogenous variables.}

\item{diag_value}{Positive numeric diagonal initial value.}
}
\value{
Dense covariance matrix (LZ x LZ).
}
\description{
Initialize a dense covariance matrix for LZ endogenous variables
}
