% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{parse_complex_formula}
\alias{parse_complex_formula}
\title{Parse complex formula specifications}
\usage{
parse_complex_formula(f)
}
\arguments{
\item{f}{A formula or formula-like object.}
}
\value{
List of formula blocks.
}
\description{
Supports formulas with '|' separators to denote additional model blocks.
For example:
\itemize{
\item y ~ x1 + x2
\item y ~ x1 + x2 | tvg.dummy ~ q1 + q2
\item tvg.dummy ~ 1 + z1 + z2 | z1 ~ q1 + q2 | z2 ~ q3 + q4
}
}
