% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-math.R
\name{riwish_safe}
\alias{riwish_safe}
\title{Safe inverse-Wishart draw wrapper}
\usage{
riwish_safe(nu, S)
}
\arguments{
\item{nu}{Degrees of freedom.}

\item{S}{Scale matrix.}
}
\value{
Sample from inverse-Wishart.
}
\description{
Uses MCMCpack::riwish if available; otherwise errors with a clear message.
}
