% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-panel.R
\name{treated_ids_from_panel}
\alias{treated_ids_from_panel}
\title{Compute treated unit ids from treatment column}
\usage{
treated_ids_from_panel(
  dta,
  id_col,
  tr_col,
  treat_type = c("binary", "continuous"),
  treat_threshold = 0
)
}
\arguments{
\item{dta}{data.table}

\item{id_col}{name of unit id column}

\item{tr_col}{name of treatment column}

\item{treat_type}{"binary" or "continuous"}

\item{treat_threshold}{threshold to define treated when continuous (default >0)}
}
\value{
vector of treated ids
}
\description{
Compute treated unit ids from treatment column
}
