% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-weights.R
\name{treated_ids_from_W}
\alias{treated_ids_from_W}
\title{Infer treated ids from a weights matrix}
\usage{
treated_ids_from_W(W)
}
\arguments{
\item{W}{Canonical weights matrix.}
}
\value{
Character vector of treated unit ids (colnames).
}
\description{
Infer treated ids from a weights matrix
}
