% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-weights.R
\name{weights_from_multisynth}
\alias{weights_from_multisynth}
\title{Coerce augsynth::multisynth object into canonical weights matrix}
\usage{
weights_from_multisynth(msynth)
}
\arguments{
\item{msynth}{multisynth object.}
}
\value{
Canonical weights matrix.
}
\description{
multisynth documents that each column corresponds to a treated unit.
In implementation, rownames are set to unit IDs.
}
