% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampling.R
\name{gibbs_postscm}
\alias{gibbs_postscm}
\title{Gibbs sampling dispatcher for post-SCM models}
\usage{
gibbs_postscm(
  gdata,
  n_iter = 1000,
  burn_in = 500,
  run_selection_gibbs = FALSE,
  do.CF = TRUE,
  selection.method = c("2SRI", "GR"),
  Z_cov_dense = FALSE,
  CF.interactions = TRUE
)
}
\arguments{
\item{gdata}{Output from prepare_data_general().}

\item{n_iter}{Number of iterations.}

\item{burn_in}{Burn-in iterations.}

\item{run_selection_gibbs}{Logical; run selection equation Gibbs block (only relevant for full sampler).}

\item{do.CF}{Use control function when IV blocks exist.}

\item{selection.method}{Character; "2SRI" or "GR" (only relevant if run_selection_gibbs=TRUE).}

\item{Z_cov_dense}{Dense covariance for IV first stage (requires MCMCpack if TRUE).}

\item{CF.interactions}{Expand CF residual interactions.}
}
\value{
Posterior draws.
}
\description{
Runs an appropriate Gibbs sampler depending on the complexity level encoded in gdata$cov.
}
