% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampling.R
\name{gibbs_sampler_one_draw}
\alias{gibbs_sampler_one_draw}
\title{Single Gibbs draw for OLS model}
\usage{
gibbs_sampler_one_draw(y, X, beta, sigma2)
}
\arguments{
\item{y}{Response vector \eqn{\lbrack n \times 1 \rbrack}.}

\item{X}{Design matrix \eqn{\lbrack n \times p \rbrack}, including intercept.}

\item{beta}{Current coefficient vector, length \eqn{p}.}

\item{sigma2}{Current error variance.}
}
\value{
A list with elements \code{betaZ}, \code{sigmaZ2},
and \code{residuals}.
}
\description{
Performs one iteration of a Gibbs sampler for a simple OLS regression
model under a flat prior on \eqn{\beta} and an improper Jeffreys prior
on \eqn{\sigma^2}.
}
