% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampling.R
\name{gibbs_sampling_moderators}
\alias{gibbs_sampling_moderators}
\title{Gibbs sampler for single-outcome post-SCM model with moderators}
\usage{
gibbs_sampling_moderators(gdata, n_iter = 1000, burn_in = 500)
}
\description{
Internal sampler for the generalized scmBayesPost object when a second-stage
moderator equation is present, but no IV and no selection equation are used.
}
\keyword{internal}
