% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampling.R
\name{gibbs_sampling_selection_moderators}
\alias{gibbs_sampling_selection_moderators}
\title{Gibbs sampler — Bayesian probit first stage + second-stage moderators}
\usage{
gibbs_sampling_selection_moderators(
  gdata,
  n_iter = 1000,
  burn_in = 500,
  control = NULL
)
}
\description{
Gibbs sampler — Bayesian probit first stage + second-stage moderators
}
\keyword{internal}
