% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{scmBayesPost-package}
\alias{scmBayesPost}
\alias{scmBayesPost-package}
\title{scmBayesPost: Post-Synthetic-Control Hierarchical Modeling with General Weight Inputs}
\description{
Generalizes mscPost to accept synthetic-control / augmented-synth weights from multiple sources.
}
\author{
\strong{Maintainer}: Andre Bonfrer \email{andre.bonfrer@deakin.edu.au}

Authors:
\itemize{
  \item Andre Bonfrer \email{andre.bonfrer@deakin.edu.au}
}

}
\keyword{internal}
